Move to 17th version of dpdk library
* Fix deprecated dpdk16 fun
* segment walker and checker added for debugging
* Fix a bug that in certain condition use to create a loop in mbufchain
* MEMPOOL single consumer actived
* Now if it's compiled with ring support, mem mbuf pool is multiple producer
* txconf adjust
* Add exception tcp-data-setup-rate for 1024 bit testcase
* Fix checksum offload in vmxnet
Since mbuf->l3_len is used for checksum offload for ipv4 but even for
layer 4 checksum, we have to set it even if we don't compute ipv4 checksum.
We use to set "NO CHECKSUM OFFLOAD" flags even without SW CHECKSUMMING

* Fix a wrong initialization
http://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf
If an int can represent all values of the original type, the value is
converted to an int; otherwise, it is converted to an unsigned int.
These are called the integer promotions.48) All other types are unchanged
by the integer promotions.

Perf test results for jspg3: 
2018-07-24 09:43:41,603 - TestPerf - INFO - Test test_01_4M_tcp_sess_setup_rate
2018-07-24 09:45:05,700 - TestPerf - INFO - Average Rate 9199481
2018-07-24 09:45:05,701 - TestPerf - INFO - Test test_02_8M_tcp_sess_setup_rate
2018-07-24 09:47:36,729 - TestPerf - INFO - Average Rate 8944980
2018-07-24 09:47:36,729 - TestPerf - INFO - Test test_03_10M_tcp_sess_setup_rate
2018-07-24 09:50:41,628 - TestPerf - INFO - Average Rate 8857241
2018-07-24 09:50:41,628 - TestPerf - INFO - Test test_04_4M_tcp_sess_data_10b_setup_rate
2018-07-24 09:52:17,462 - TestPerf - INFO - Average Rate 3735844
2018-07-24 09:52:17,462 - TestPerf - INFO - Test test_05_4M_tcp_sess_data_1024b_setup_rate
2018-07-24 09:53:52,714 - TestPerf - INFO - Average Rate 2717411
2018-07-24 09:53:52,714 - TestPerf - INFO - Test test_06_4M_tcp_sess_data_1300b_setup_rate
2018-07-24 09:55:28,573 - TestPerf - INFO - Average Rate 3901975
2018-07-24 09:55:28,573 - TestPerf - INFO - Test test_07_4M_udp_sess_data_10b_setup_rate
2018-07-24 09:56:49,809 - TestPerf - INFO - Average Rate 11498273
2018-07-24 09:56:49,810 - TestPerf - INFO - Test test_08_4M_http_sess_data_10b_setup_rate
2018-07-24 09:58:23,962 - TestPerf - INFO - Average Rate 3371774
2018-07-24 09:58:23,962 - TestPerf - INFO - Test test_09_4M_udp_mcast_flows_data_10b_setup_rate
2018-07-24 09:59:46,752 - TestPerf - INFO - Average Rate 22278973
2018-07-24 09:59:46,752 - TestPerf - INFO - Test test_10_timestamp_4M_tcp_sess_setup_rate
2018-07-24 10:01:18,559 - TestPerf - INFO - Average Rate 8056233
2018-07-24 10:01:18,559 - TestPerf - INFO - Test test_11_timestamp_4M_udp_sess_data_10b_setup_rate
2018-07-24 10:02:42,626 - TestPerf - INFO - Average Rate 10793233
2018-07-24 10:02:42,626 - TestPerf - INFO - Test test_12_recent_timestamp_4M_tcp_sess_setup_rate
2018-07-24 10:04:13,953 - TestPerf - INFO - Average Rate 7944913
2018-07-24 10:04:13,953 - TestPerf - INFO - Test test_13_recent_timestamp_4M_udp_sess_data_10b_setup_rate
2018-07-24 10:05:37,850 - TestPerf - INFO - Average Rate 10771231
2018-07-24 10:05:37,850 - TestPerf - INFO - Test test_14_timestamp_raw_4M_tcp_sess_data_1024b_setup_rate
2018-07-24 10:07:13,146 - TestPerf - INFO - Average Rate 3097852
2018-07-24 10:07:13,146 - TestPerf - INFO - Test test_15_timestamp_raw_4M_udp_sess_data_1024b_setup_rate
2018-07-24 10:08:30,078 - TestPerf - INFO - Average Rate 13117234
2018-07-24 10:08:30,078 - TestPerf - INFO - Test test_16_recent_timestamp_raw_4M_tcp_sess_data_1024b_setup_rate
2018-07-24 10:10:05,307 - TestPerf - INFO - Average Rate 3056690
2018-07-24 10:10:05,307 - TestPerf - INFO - Test test_17_recent_timestamp_raw_4M_udp_sess_data_1024b_setup_rate
2018-07-24 10:11:22,256 - TestPerf - INFO - Average Rate 13109466

+++ src/kni_if/tpg_kni_if.c
@@ -64,8 +64,8 @@
 /*****************************************************************************
  * Definitions
  ****************************************************************************/
-static int kni_change_mtu(uint8_t port, unsigned mtu);
-static int kni_config_network_if(uint8_t port, uint8_t if_state);
+static int kni_change_mtu(uint16_t port, unsigned mtu);
+static int kni_config_network_if(uint16_t port, uint8_t if_state);
 
 /*****************************************************************************
  * Globals
@@ -275,7 +275,7 @@ bool kni_if_init(void)
 /*****************************************************************************
  * kni_change_mtu()
  ****************************************************************************/
-static int kni_change_mtu(uint8_t port, unsigned mtu)
+static int kni_change_mtu(uint16_t port, unsigned mtu)
 {
     global_config_t *cfg;
 
@@ -325,7 +325,7 @@ static int kni_change_mtu(uint8_t port, unsigned mtu)
 /*****************************************************************************
  * kni_config_network_if()
  ****************************************************************************/
-static int kni_config_network_if(uint8_t port, uint8_t if_state)
+static int kni_config_network_if(uint16_t port, uint8_t if_state)
 {
     RTE_LOG(DEBUG, USER1, "KNI: Configure network interface of %d %s\n",
             port, if_state ? "up" : "down");
